Apply stashed changes after merging main
